backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic>=2.0.0
authlib
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
starlette
itsdangerous
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
python-dotenv
redis
celery
boto3
python-magic
PyPDF2